+++ gems/aws-sdk-s3/CHANGELOG.md
@@ -1,7 +1,10 @@
 Unreleased Changes
 ------------------
 
+* Issue - Fix FIPS and global endpoint behavior for S3 ARNs.
+
 * Issue - Increases `multipart_threshold` default from 15 megabytes to 100 megabytes.
+
 
 1.93.0 (2021-03-24)
 ------------------